content-api/requirements.txt
FastAPI
uvicorn
supabase
pydantic
python-dotenv
sqlalchemy